+++ psx/input/sda.c
@@ -122,7 +122,7 @@ void psxi_sda_on_button_release(void* udata, uint32_t data) {
 }
 
 // To-do: Implement analog mode
-void psxi_sda_on_analog_change(void* udata, uint32_t axis, uint8_t data) {
+void psxi_sda_on_analog_change(void* udata, uint32_t axis, uint16_t data) {
     // Suppress warning until we implement analog mode
     psxi_sda_t* sda = (psxi_sda_t*)udata;
